requirements.txt
streamlit
langchain-core
langchain-community
chromadb
zhipuai
python-dotenv